+++ R/inbo_2015_colours.R
@@ -1,4 +1,4 @@
-#' A colour pallet according the INBO style guide version >= 2015.
+#' A Colour Palette According the INBO Style Guide Version >= 2015.
 #'
 #' @export
 #' @param n The number of colours
@@ -21,4 +21,30 @@ inbo_palette <- function(n) {
     )
   }
   rep(palette, length.out = n)
+}
+
+#' A Colour Palette for NARA reports.
+#'
+#' @export
+#' @inheritParams inbo_palette
+#' @return a vector of n hexadecimal values defining the colours.
+#' @family colours
+#' @importFrom assertthat assert_that is.count noNA
+nara_palette <- function(n) {
+  palette <- c(
+    inbo_hoofd, inbo_steun_geelgroen, inbo_steun_blauw, inbo_oranje,
+    inbo_steun_donkerroos, inbo_groen, inbo_grijsblauw
+  )
+  if (missing(n)) {
+    n <- length(palette)
+  }
+  assert_that(is.count(n), noNA(n))
+  assert_that(
+    n <= length(palette),
+    msg = sprintf(
+      "`nara_palette()` has only %i colours, you requested %i.",
+      length(palette), n
+    )
+  )
+  palette[seq_len(n)]
 }